+++ lib/vdot_handler.py
@@ -18,22 +18,11 @@ class VDOTHandler:
             # Last resort: assume last column
             self.marathon_col = self.vdot_df.columns[-1]
 
-    def get_vdot_options(self):
-        """Returns list of available VDOTs for dropdown"""
-        return sorted(self.vdot_df['VDOT'].astype(float).unique())
-
     def get_time_for_vdot(self, vdot):
         """Returns marathon time string for a given VDOT"""
         row = self.vdot_df[self.vdot_df['VDOT'] == vdot]
         if not row.empty:
             return row.iloc[0][self.marathon_col]
-        return None
-
-    def get_seconds_for_vdot(self, vdot):
-        """Returns marathon time in seconds for a given VDOT"""
-        t_str = self.get_time_for_vdot(vdot)
-        if t_str:
-            return self._time_str_to_seconds(t_str)
         return None
 
     def get_closest_vdot(self, target_time_str):
@@ -56,7 +45,7 @@ class VDOTHandler:
              seconds = self.vdot_df[self.marathon_col].apply(self._time_str_to_seconds).values
              vdots = self.vdot_df['VDOT'].astype(float).values
              return vdots, seconds
-        except:
+        except (KeyError, ValueError, TypeError):
             return np.array([]), np.array([])
             
     def get_time_for_exact_vdot(self, vdot_float):
@@ -100,5 +89,5 @@ class VDOTHandler:
             elif len(parts) == 2: # m:ss (unlikely for marathon but possible for pace)
                 return int(parts[0])*60 + int(parts[1])
             return 0
-        except:
+        except (ValueError, TypeError):
             return 0